add duration of answer in test and fix bug with liked question

+++ src/main/java/com/driving/school/model/StudentAnswersTest.java
@@ -23,7 +23,7 @@ public class StudentAnswersTest {
     private Boolean skipped;
 
     @Column(name = "DURATIONOFANSWER", nullable = false)
-    private Integer durationOfAnswer;
+    private Long durationOfAnswer;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @OnDelete(action = OnDeleteAction.CASCADE)
